requirements/production.txt
Flask
Flask-Sqlalchemy
psycopg2-binary
Flask-Migrate
pymemcache
gunicorn
